feat(healthy-food-delivery): code splitting & webpack
affects: @practice/healthy-food-delivery

splitting by modules which bundle via webpack
make functions more SRP

ISSUES CLOSED: #58
