+++ practice/healthy-food-delivery/package.json
@@ -4,7 +4,15 @@
   "license": "BSD-3-Clause",
   "main": "src/index.ts",
   "scripts": {
-    "build": "tsc",
+    "build": "webpack",
+    "compile": "tsc",
     "lint": "eslint"
+  },
+  "devDependencies": {
+    "eslint-import-resolver-typescript": "^2.4.0",
+    "ts-loader": "^8.0.17",
+    "ts-node": "^9.1.1",
+    "webpack": "^5.22.0",
+    "webpack-cli": "^4.5.0"
   }
 }
