+++ practice/healthy-food-delivery/public/index.html
@@ -282,7 +282,7 @@
     </div>
   </div>
 
-  <script src="../dist/index.js"></script>
+  <script src="../dist/bundle.js"></script>
 </body>
 
 </html>
