--- practice/healthy-food-delivery/src/db.json
@@ -1,37 +1,0 @@
-{
-  "menu": [
-    {
-      "imgSrc": "../src/assets/img/tabs/vegy.jpg",
-      "alt": "fitness",
-      "title": "Меню 'Фитнес'",
-      "description": "Меню 'Фитнес' - это новый подход к приготовлению блюд: больше свежих овощей и фруктов. Продукт активных и здоровых людей. Это абсолютно новый продукт с оптимальной ценой и высоким качеством!",
-      "price": 9
-    },
-    {
-      "imgSrc": "../src/assets/img/tabs/elite.jpg",
-      "alt": "post",
-      "title": "Меню 'Постное'",
-      "description": "Меню 'Постное' - это тщательный подбор ингредиентов: полное отсутствие продуктов животного происхождения, молоко из миндаля, овса, кокоса или гречки, правильное количество белков за счет тофу и импортных вегетарианских стейков.",
-      "price": 14
-    },
-    {
-      "imgSrc": "../src/assets/img/tabs/elite.jpg",
-      "alt": "premium",
-      "title": "Меню 'Премиум'",
-      "description": "В меню 'Премиум' мы используем не только красивый дизайн упаковки, но и качественное исполнение блюд. Красная рыба, морепродукты, фрукты - ресторанное меню без похода в ресторан!",
-      "price": 21
-    }
-  ],
-  "requests": [
-    {
-      "name": "Vladislav Bukatov",
-      "phone": "+79616506174",
-      "id": 1
-    },
-    {
-      "name": "themetronome",
-      "phone": "+79658714587",
-      "id": 2
-    }
-  ]
-}
