+++ practice/healthy-food-delivery/src/services/postData.ts
@@ -1,0 +1,11 @@
+export async function sendFormData(url: string, data: string) {
+  const result: Response = await fetch(url, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: data,
+  })
+
+  return result.json()
+}
