+++ practice/healthy-food-delivery/src/utils/modalWindowTimer.ts
@@ -1,0 +1,6 @@
+import { openModalWindow } from '../modules/userDataModal'
+
+export const modalWindowTimer: number = window.setTimeout(
+  () => openModalWindow('.modal', modalWindowTimer),
+  3000
+)
